feat: update benchmark

+++ sqlite-bench/util.c
@@ -23,11 +23,11 @@ bool starts_with(const char* str, const char* pre) {
 
 char* trim_space(const char* s) {
   size_t start = 0;
-  while (start < strlen(s) && isspace(s[start])) {
+  while (start < strlen(s) && isspace((int)s[start])) {
     start++;
   }
   size_t limit = strlen(s);
-  while (limit > start && isspace(s[limit - 1])) {
+  while (limit > start && isspace((int)s[limit - 1])) {
     limit--;
   }
